requirements.txt
streamlit
fpdf
feedparser
python-dotenv
